requirements.txt
Django~=5.0.3
django-environ~=0.11.2
django-nested-inline~=0.4.6
django-tinymce~=3.7.1
djangorestframework~=3.14.0
factory-boy~=3.3.0
psycopg~=3.1.8
